:whale:文档更新:快速从零编写服务器完整示例 DebugInOut     打印微调     打印用户自定义异常的msg
新增 快速从零编写服务器完整示例代码
    HelloReq
    DemoAction
    DemoLogicServer

    DemoApplication
    DemoWebsocketClient

BarSkeletonBuilderParamConfig 新增构建 BarSkeletonBuilder 方法

+++ game-collect/game-test/src/main/java/com/iohao/game/collect/external/tester/module/OnMessage.java
@@ -16,8 +16,8 @@ public interface OnMessage {
     default ExternalMessage createExternalMessage() {
 
         ExternalMessage request = new ExternalMessage();
-        request.setCmdCode((short) 1);
-        request.setProtocolSwitch((byte) 0);
+        request.setCmdCode( 1);
+        request.setProtocolSwitch(0);
 
         request.setCmdMerge(this.getCmdMerge());
